+++ ExpressionEval/Nodes/MultiplyNode.cs
@@ -3,6 +3,6 @@
 		public MultiplyNode() {}
 		public MultiplyNode(Node left, Node right) : base(left, right) {}
 
-		public override int Value => Left * Right;
+		public override float Value => Left * Right;
 	}
 }
