+++ ExpressionEval/Nodes/SubtractNode.cs
@@ -3,6 +3,6 @@
 		public SubtractNode() {}
 		public SubtractNode(Node left, Node right) : base(left, right) {}
 
-		public override int Value => Left - Right;
+		public override float Value => Left - Right;
 	}
 }
